fix: Resolve the hierarchy issue of toast and dialog
